Extract `parseDiff` from `diff-to-document` helper into a separate util `parse-diff-as-document`
